app-server: split unix websocket ctrl-c tests
